refactor sqlite federation and chirp error handling
