+++ be1-go/internal/sqlite/sqlite.go
@@ -6,10 +6,8 @@ import (
 	"encoding/json"
 	"errors"
 	"go.dedis.ch/kyber/v3"
-	"golang.org/x/xerrors"
 	_ "modernc.org/sqlite"
 	"popstellar/internal/crypto"
-	"popstellar/internal/message/messagedata"
 	"popstellar/internal/message/query/method/message"
 	"strings"
 	"time"
@@ -312,14 +310,6 @@ func (s *SQLite) HasMessage(messageID string) (bool, error) {
 	}
 }
 
-//======================================================================================================================
-// RootRepository interface implementation
-//======================================================================================================================
-
-//======================================================================================================================
-// LaoRepository interface implementation
-//======================================================================================================================
-
 func (s *SQLite) GetOrganizerPubKey(laoPath string) (kyber.Point, error) {
 	dbLock.Lock()
 	defer dbLock.Unlock()
@@ -335,200 +325,4 @@ func (s *SQLite) GetOrganizerPubKey(laoPath string) (kyber.Point, error) {
 		return nil, err
 	}
 	return organizerPubKey, nil
-}
-
-//======================================================================================================================
-// ElectionRepository interface implementation
-//======================================================================================================================
-
-//======================================================================================================================
-// ChirpRepository interface implementation
-//======================================================================================================================
-
-func (s *SQLite) StoreChirpMessages(channel, generalChannel string, msg, generalMsg message.Message) error {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	msgBytes, err := json.Marshal(msg)
-	if err != nil {
-		return err
-	}
-	messageData, err := base64.URLEncoding.DecodeString(msg.Data)
-	if err != nil {
-		return err
-	}
-	generalMsgBytes, err := json.Marshal(generalMsg)
-	if err != nil {
-		return err
-	}
-	generalMessageData, err := base64.URLEncoding.DecodeString(generalMsg.Data)
-	if err != nil {
-		return err
-	}
-	storedTime := time.Now().UnixNano()
-
-	err = s.insertMessageHelper(tx, msg.MessageID, msgBytes, messageData, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, channel, msg.MessageID, true)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertMessage, generalMsg.MessageID, generalMsgBytes, generalMessageData, storedTime)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(insertChannelMessage, generalChannel, generalMsg.MessageID, false)
-	if err != nil {
-		return err
-	}
-	err = tx.Commit()
-	return err
-}
-
-//======================================================================================================================
-// ReactionRepository interface implementation
-//======================================================================================================================
-
-//======================================================================================================================
-// FederationRepository interface implementation
-//======================================================================================================================
-
-func (s *SQLite) IsChallengeValid(senderPk string, challenge messagedata.FederationChallenge, channelPath string) error {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	var federationChallengeBytes []byte
-	err := s.database.QueryRow(selectValidFederationChallenges, channelPath,
-		senderPk, messagedata.FederationObject,
-		messagedata.FederationActionChallenge, challenge.Value,
-		challenge.ValidUntil).Scan(&federationChallengeBytes)
-	if err != nil {
-		return err
-	}
-
-	var federationChallenge messagedata.FederationChallenge
-	err = json.Unmarshal(federationChallengeBytes, &federationChallenge)
-	if err != nil {
-		return err
-	}
-
-	if federationChallenge != challenge {
-		return xerrors.New("the federation challenge doesn't match")
-	}
-
-	return nil
-}
-
-func (s *SQLite) RemoveChallenge(challenge messagedata.FederationChallenge) error {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	result, err := s.database.Exec(deleteFederationChallenge,
-		messagedata.FederationObject,
-		messagedata.FederationActionChallenge, challenge.Value,
-		challenge.ValidUntil)
-	if err != nil {
-		return err
-	}
-
-	nb, err := result.RowsAffected()
-	if err != nil {
-		return err
-	}
-
-	if nb != 1 {
-		return xerrors.New("unexpected number of rows affected")
-	}
-
-	return nil
-}
-
-func (s *SQLite) GetFederationExpect(senderPk string, remotePk string, challenge messagedata.FederationChallenge, channelPath string) (messagedata.FederationExpect, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	rows, err := s.database.Query(selectFederationExpects, channelPath,
-		senderPk, messagedata.FederationObject,
-		messagedata.FederationActionExpect, remotePk)
-	if err != nil {
-		return messagedata.FederationExpect{}, err
-	}
-	defer rows.Close()
-
-	// iterate over all FederationExpect sent from the given sender pk,
-	// and search the one matching the given FederationChallenge
-	for rows.Next() {
-		var federationExpectBytes []byte
-		err = rows.Scan(&federationExpectBytes)
-		if err != nil {
-			continue
-		}
-
-		var federationExpect messagedata.FederationExpect
-		err = json.Unmarshal(federationExpectBytes, &federationExpect)
-		if err != nil {
-			continue
-		}
-
-		var federationChallenge messagedata.FederationChallenge
-		errAnswer := federationExpect.ChallengeMsg.UnmarshalMsgData(&federationChallenge)
-		if errAnswer != nil {
-			return messagedata.FederationExpect{}, errAnswer
-		}
-
-		if federationChallenge == challenge {
-			return federationExpect, nil
-		}
-	}
-
-	return messagedata.FederationExpect{}, sql.ErrNoRows
-}
-
-func (s *SQLite) GetFederationInit(senderPk string, remotePk string, challenge messagedata.FederationChallenge, channelPath string) (messagedata.FederationInit, error) {
-	dbLock.Lock()
-	defer dbLock.Unlock()
-
-	rows, err := s.database.Query(selectFederationExpects, channelPath,
-		senderPk, messagedata.FederationObject,
-		messagedata.FederationActionInit, remotePk)
-	if err != nil {
-		return messagedata.FederationInit{}, err
-	}
-	defer rows.Close()
-
-	// iterate over all FederationInit sent from the given sender pk,
-	// and search the one matching the given FederationChallenge
-	for rows.Next() {
-		var federationInitBytes []byte
-		err = rows.Scan(&federationInitBytes)
-		if err != nil {
-			continue
-		}
-
-		var federationInit messagedata.FederationInit
-		err = json.Unmarshal(federationInitBytes, &federationInit)
-		if err != nil {
-			continue
-		}
-
-		var federationChallenge messagedata.FederationChallenge
-		errAnswer := federationInit.ChallengeMsg.UnmarshalMsgData(&federationChallenge)
-		if errAnswer != nil {
-			return messagedata.FederationInit{}, errAnswer
-		}
-
-		if federationChallenge == challenge {
-			return federationInit, nil
-		}
-	}
-
-	return messagedata.FederationInit{}, sql.ErrNoRows
 }
